+++ src/main/java/com/wojcik/lukasz/melanomacheckerserver/model/detector/DetectionAware.java
@@ -2,5 +2,5 @@ package com.wojcik.lukasz.melanomacheckerserver.model.detector;
 
 public interface DetectionAware<T> {
 
-    Float detect(T parameter);
+    Integer detect(T parameter);
 }
